# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "DCO Mart API",
        "contact": {}
    },
    "paths": {
        "/category": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Get All Categories",
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Create Category",
                "parameters": [
                    {
                        "description": "Create Category Data",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCategoryDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/category/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Get Category By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Update Category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create Category Data",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCategoryDTO"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Delete Category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/item": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Item"
                ],
                "summary": "Get All Item",
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Item"
                ],
                "summary": "Create Item",
                "parameters": [
                    {
                        "description": "Create Item Data",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateItemDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/item/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Item"
                ],
                "summary": "Get Item By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Item"
                ],
                "summary": "Update Item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Item Data",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateItemDTO"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Item"
                ],
                "summary": "Delete Item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/order": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Get All Order",
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Create Order",
                "parameters": [
                    {
                        "description": "Create Order Data",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrderDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/order/checkout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Checkout Order",
                "parameters": [
                    {
                        "description": "Order payload",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrderDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/order/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Get Order By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/order/{id}/method": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Update Method Payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Method Payment",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateMethodDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/order/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Update Status Order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Status Order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateStatusDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login User",
                "parameters": [
                    {
                        "description": "User Login",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Register User",
                "parameters": [
                    {
                        "description": "User Registration Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterDTO"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "dto.CreateCategoryDTO": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Example Category"
                }
            }
        },
        "dto.CreateItemDTO": {
            "type": "object",
            "required": [
                "category_id",
                "desc",
                "name",
                "photo",
                "price"
            ],
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "desc": {
                    "type": "string",
                    "example": "Example Desc Item"
                },
                "name": {
                    "type": "string",
                    "example": "Example Item"
                },
                "photo": {
                    "description": "Example: a byte array",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "price": {
                    "type": "number",
                    "example": 1
                }
            }
        },
        "dto.CreateOrderDTO": {
            "type": "object",
            "required": [
                "method",
                "order_detail",
                "status",
                "user_id"
            ],
            "properties": {
                "method": {
                    "type": "string",
                    "example": "transfer"
                },
                "order_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CreateOrderDetailDTO"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "pending"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.CreateOrderDetailDTO": {
            "type": "object",
            "required": [
                "item_id",
                "order_id",
                "quantity"
            ],
            "properties": {
                "item_id": {
                    "type": "integer",
                    "example": 1
                },
                "order_id": {
                    "type": "integer",
                    "example": 1
                },
                "quantity": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "dto.LoginDTO": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "admin@example.com"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "admin123"
                }
            }
        },
        "dto.RegisterDTO": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "phone"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "123 Main St"
                },
                "email": {
                    "type": "string",
                    "example": "admin@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "Admin"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "admin123"
                },
                "phone": {
                    "type": "string",
                    "example": "1234567890"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "dto.UpdateItemDTO": {
            "type": "object",
            "required": [
                "category_id",
                "desc",
                "name",
                "photo",
                "price"
            ],
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 2
                },
                "desc": {
                    "type": "string",
                    "example": "Example Update Desc Item"
                },
                "name": {
                    "type": "string",
                    "example": "Example Update Item"
                },
                "photo": {
                    "description": "Example: a byte array",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "price": {
                    "type": "number",
                    "example": 2
                }
            }
        },
        "dto.UpdateMethodDTO": {
            "type": "object",
            "required": [
                "method"
            ],
            "properties": {
                "method": {
                    "type": "string",
                    "example": "cod"
                }
            }
        },
        "dto.UpdateStatusDTO": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "example": "delivered"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}